disassembler blacklist
